requirements.txt
langchain
openai
fastapi
uvicorn
pydantic
pytest
python-dotenv
streamlit
langchain-groq
groq